test claw wifi bork

+++ Managment/ReignOS.Service/HardwarePatches/WiFiPatches.cs
@@ -52,14 +52,14 @@ esac";
             string path = "/usr/lib/systemd/system-sleep";
             if (!Directory.Exists(path)) Directory.CreateDirectory(path);
             path = Path.Combine(path, "wifi-sleep.sh");
-            /*if (File.Exists(path))
+            if (File.Exists(path))
             {
                 if (apply) return;// already patched
 
                 // remove
                 File.Delete(path);
                 return;
-            }*/
+            }
 
             const string config =
 @"#!/bin/bash
@@ -84,14 +84,14 @@ esac";
             string path = "/usr/lib/systemd/system-sleep";
             if (!Directory.Exists(path)) Directory.CreateDirectory(path);
             path = Path.Combine(path, "wifi-sleep.sh");
-            /*if (File.Exists(path))
+            if (File.Exists(path))
             {
                 if (apply) return;// already patched
 
                 // remove
                 File.Delete(path);
                 return;
-            }*/
+            }
 
             const string config =
 @"#!/bin/bash
